requirements.txt
streamlit==1.31.1
pandas==2.2.0
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
gspread==5.12.4
oauth2client==4.1.3
openai==1.12.0
plotly==5.18.0
openpyxl==3.1.2
requests==2.31.0
httpx==0.26.0 